feat: 添加一些demo 1.增加 showcase demo 2.增加 多种Recycleview 展示demo

+++ app/src/main/java/com/apache/fastandroid/demo/drawable/DrawableView.kt
@@ -19,7 +19,7 @@ class DrawableView(context: Context?, attrs: AttributeSet?) : View(context, attr
 
     override fun onDraw(canvas: Canvas) {
         super.onDraw(canvas)
-        drawable.setBounds(20.dp.toInt(),40.dp.toInt(),width,height)
+        drawable.setBounds(20.dp.toInt(), 40.dp,width,height)
         drawable.draw(canvas)
     }
 }
